fix: clear line on new entry and correct hardcoded vals
When entering a new value onto the stack the first char on the entry
line was correctly 0 but the second char wasn't overwritten with a
space. Manually corrected that by placing a space there. Previously, the
CEDev kit had incorrect names for sin and cos. That's been corrected so
there's no reason to keep the buttons switched. Pi is correct as well
now.